LUNAPARK-1466 - fail-lock parameter

+++ Tank/Core.py
@@ -37,6 +37,8 @@ class TankCore:
 
         self.artifacts_base_dir = self.get_option(self.SECTION, "artifacts_base_dir", tempfile.gettempdir())
         self.artifacts_dir = self.get_option(self.SECTION, "artifacts_dir", "")
+        if self.artifacts_dir:
+            self.artifacts_dir=os.path.expanduser(self.artifacts_dir)
 
         for (plugin_name, plugin_path) in self.config.get_options(self.SECTION, self.PLUGIN_PREFIX):
             if not plugin_path:
@@ -196,11 +198,17 @@ class TankCore:
             self.log.warning("No artifacts dir configured")
             return            
         
-        self.log.debug("Collecting file: %s to %s", filename, self.artifacts_dir + '/' + os.path.basename(filename))
+        dest=self.artifacts_dir + '/' + os.path.basename(filename)
+        self.log.debug("Collecting file: %s to %s", filename, dest)
         if not filename or not os.path.exists(filename):
             self.log.warning("File not found to collect: %s", filename)
             return
         
+        if os.path.exists(dest):
+            # FIXME: find a way to store artifacts anyway
+            self.log.warning("File already exists: %s", dest)
+            return
+                
         if keep_original:
             shutil.copy(filename, self.artifacts_dir)
         else:
